backend/requirements.txt
playwright
supabase
python-dotenv
sentry-sdk
resend
requests
